remove chunks and use context manager, also fix cast type problem

+++ lib/osop/regridders.py
@@ -47,8 +47,11 @@ def regrid_cons_masked(source_in, var, target_in, thresh=0.5):
     # now regrid LSM to get a fractionsl LSM on the target grid
     lsm_out = regridder(source["lsm"], keep_attrs=True)
 
-    # mask the source using same approach as above but integer
-    source["mask"] = xr.where(~np.isnan(source[var].isel(time=0)), 1, 0)
+    # mask the source but ensure float dtype
+    # (avoid integer dtype which causes xESMF to cast NaNs to ints)
+    source["mask"] = xr.where(~np.isnan(source[var].isel(time=0)), 1.0, 0.0).astype(
+        source[var].dtype
+    )
 
     # for the target grid, based on where the regridded mask is at thresh or more
     target["mask"] = xr.where(lsm_out > thresh, 1.0, 0.0)
